requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
aiohttp==3.9.1
beautifulsoup4==4.9.3
html5lib==1.1
requests==2.26.0
Flask-Talisman==1.0.0
Flask-SeaSurf==1.1.1
Flask-Limiter==2.8.1
python-dotenv==0.19.0
bleach==4.1.0
lxml
pytest==7.4.3
black==23.11.0
flake8==6.1.0
